fix: remove secret log from config
